split long cypress  specs
