+++ app/client/cypress/e2e/Regression/ClientSide/Binding/Promises_Spec.ts
@@ -17,46 +17,7 @@ describe("Validate basic Promises", () => {
     agHelper.SaveLocalStorageCache();
   });
 
-  it("1. Verify storeValue via .then via direct Promises", () => {
-    const date = new Date().toDateString();
-    cy.fixture("promisesBtnDsl").then((val: any) => {
-      agHelper.AddDsl(val, locator._spanButton("Submit"));
-    });
-    ee.SelectEntityByName("Button1", "Widgets");
-    propPane.EnterJSContext(
-      "onClick",
-      "{{storeValue('date', Date()).then(() => showAlert(appsmith.store.date))}}",
-    );
-    deployMode.DeployApp();
-    agHelper.ClickButton("Submit");
-    agHelper.ValidateToastMessage(date);
-    deployMode.NavigateBacktoEditor();
-  });
-
-  it("2. Verify resolve & chaining via direct Promises", () => {
-    cy.fixture("promisesBtnDsl").then((val: any) => {
-      agHelper.AddDsl(val, locator._spanButton("Submit"));
-    });
-    ee.SelectEntityByName("Button1", "Widgets");
-    propPane.EnterJSContext(
-      "onClick",
-      `{{
-          new Promise((resolve) => {
-            resolve("We are on planet")
-          }).then((res) => {
-            return res + " Earth"
-          }).then((res) => {
-            showAlert(res, 'success')
-          }).catch(err => { showAlert(err, 'error') });
-        }}`,
-    );
-    deployMode.DeployApp();
-    agHelper.ClickButton("Submit");
-    agHelper.ValidateToastMessage("We are on planet Earth");
-    deployMode.NavigateBacktoEditor();
-  });
-
-  it("3. Verify Async Await in direct Promises", () => {
+  it("1. Verify Async Await in direct Promises", () => {
     cy.fixture("promisesBtnDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
     });
@@ -94,7 +55,7 @@ describe("Validate basic Promises", () => {
     //   .contains(/male|female|null/g);
   });
 
-  it("4. Verify .then & .catch via direct Promises", () => {
+  it("2. Verify .then & .catch via direct Promises", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnImgDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
@@ -128,7 +89,7 @@ describe("Validate basic Promises", () => {
     );
   });
 
-  it("5. Verify .then & .catch via JS Objects in Promises", () => {
+  it("3. Verify .then & .catch via JS Objects in Promises", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
@@ -157,7 +118,7 @@ return InspiringQuotes.run().then((res) => { showAlert("Today's quote for " + us
       );
   });
 
-  it("6. Verify Promise.race via direct Promises", () => {
+  it("4. Verify Promise.race via direct Promises", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
@@ -185,7 +146,7 @@ return InspiringQuotes.run().then((res) => { showAlert("Today's quote for " + us
       );
   });
 
-  it("7. Verify maintaining context via direct Promises", () => {
+  it("5. Verify maintaining context via direct Promises", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnListDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
@@ -235,7 +196,7 @@ return InspiringQuotes.run().then((res) => { showAlert("Today's quote for " + us
     );
   });
 
-  it("8: Verify Promise.all via direct Promises", () => {
+  it("6: Verify Promise.all via direct Promises", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnDsl").then((val: any) => {
       agHelper.AddDsl(val, locator._spanButton("Submit"));
@@ -260,7 +221,7 @@ return InspiringQuotes.run().then((res) => { showAlert("Today's quote for " + us
     agHelper.ValidateToastMessage("cat,dog,camel,rabbit,rat");
   });
 
-  it("9. Bug 10150: Verify Promise.all via JSObjects", () => {
+  it("7. Bug 10150: Verify Promise.all via JSObjects", () => {
     deployMode.NavigateBacktoEditor();
     const date = new Date().toDateString();
     cy.fixture("promisesBtnDsl").then((val: any) => {
@@ -294,66 +255,7 @@ showAlert("Wonderful! all apis executed", "success")).catch(() => showAlert("Ple
     agHelper.AssertContains(/Wonderful|Please check/g);
   });
 
-  it("10. Verify Promises.any via direct JSObjects", () => {
-    deployMode.NavigateBacktoEditor();
-    cy.fixture("promisesBtnDsl").then((val: any) => {
-      agHelper.AddDsl(val, locator._spanButton("Submit"));
-    });
-    jsEditor.CreateJSObject(
-      `export default {
-      func2: async () => {
-     return Promise.reject(new Error('fail')).then(showAlert("Promises reject from func2"));
-    },
-      func1: async () => {
-      showAlert("In func1")
-      return "func1"
-    },
-    func3: async () => {
-      showAlert("In func3")
-      return "func3"
-    },
-    runAny: async () => {
-      return Promise.any([this.func2(), this.func3(), this.func1()]).then((value) => showAlert("Resolved promise is:" + value))
-    }
-    }`,
-      {
-        paste: true,
-        completeReplace: true,
-        toRun: false,
-        shouldCreateNewJSObj: true,
-      },
-    );
-    ee.SelectEntityByName("Button1", "Widgets");
-    cy.get("@jsObjName").then((jsObjName) => {
-      propPane.EnterJSContext("onClick", "{{" + jsObjName + ".runAny()}}");
-    });
-    deployMode.DeployApp();
-    agHelper.ClickButton("Submit");
-    agHelper.AssertElementLength(locator._toastMsg, 4);
-    agHelper.ValidateToastMessage("Promises reject from func2", 0);
-    agHelper.ValidateToastMessage("Resolved promise is:func3", 3); //Validating last index
-  });
-
-  it("11. Bug : 11110 - Verify resetWidget via .then direct Promises", () => {
-    deployMode.NavigateBacktoEditor();
-    cy.fixture("promisesBtnDsl").then((dsl: any) => {
-      agHelper.AddDsl(dsl, locator._spanButton("Submit"));
-    });
-    ee.SelectEntityByName("Button1", "Widgets");
-    propPane.EnterJSContext(
-      "onClick",
-      "{{resetWidget('Input1').then(() => showAlert(Input1.text))}}",
-    );
-    deployMode.DeployApp(locator._widgetInputSelector("inputwidgetv2"));
-    agHelper.TypeText(
-      locator._widgetInputSelector("inputwidgetv2"),
-      "Update value",
-    );
-    agHelper.ClickButton("Submit");
-    agHelper.ValidateToastMessage("Test");
-  });
-
-  it("12. Bug 9782: Verify .then & .catch (show alert should trigger) via JS Objects without return keyword", () => {
+  it("8. Bug 9782: Verify .then & .catch (show alert should trigger) via JS Objects without return keyword", () => {
     deployMode.NavigateBacktoEditor();
     cy.fixture("promisesBtnDsl").then((val: any) => {
       agHelper.AddDsl(val);
