+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/JSONForm/JSONForm_HiddenFields_2_spec.js
@@ -84,74 +84,77 @@ describe("JSON Form Hidden fields", () => {
     cy.testJsontext("text", "{{JSON.stringify(JSONForm1.formData)}}");
   });
 
-  it("1. can hide Array Field", () => {
-    cy.openPropertyPane("jsonformwidget");
-    cy.openFieldConfiguration("education");
-    hideAndVerifyProperties("education", [
-      {
-        college: "MIT",
-        year: "20/10/2014",
-      },
-    ]);
-  });
-
-  it("2. can hide Array Field's inner fields", () => {
-    cy.openPropertyPane("jsonformwidget");
-    cy.openFieldConfiguration("education");
-    cy.openFieldConfiguration("__array_item__", false);
-    cy.openFieldConfiguration("college", false);
-
-    hideAndVerifyProperties("education-0--college", "MIT", (formData) => {
-      return formData.education[0].college;
-    });
-  });
-
-  it("3. can hide Checkbox Field", () => {
+  it("1. can hide Phone Number Input Field", () => {
+    const defaultValue = "1000";
     // Add new custom field
-    addCustomField("Checkbox");
+    addCustomField("Phone Number Input");
+
+    cy.testJsontext("defaultvalue", defaultValue);
+
+    hideAndVerifyProperties("customField1", defaultValue);
+
+    removeCustomField();
+  });
+
+  it("2. can hide Radio Group Field", () => {
+    const defaultValue = "Y";
+    // Add new custom field
+    addCustomField("Phone Number Input");
+
+    cy.testJsontext("defaultvalue", defaultValue);
+
+    hideAndVerifyProperties("customField1", defaultValue);
+
+    removeCustomField();
+  });
+
+  it("3. can hide Select Field", () => {
+    const defaultValue = "BLUE";
+    // Add new custom field
+    addCustomField(/^Select/);
+
+    cy.testJsontext("defaultselectedvalue", defaultValue);
+
+    hideAndVerifyProperties("customField1", defaultValue);
+
+    removeCustomField();
+  });
+
+  it("4. can hide Switch Field", () => {
+    // Add new custom field
+    addCustomField("Switch");
 
     hideAndVerifyProperties("customField1", true);
-    // Remove custom field
+
     removeCustomField();
   });
 
-  it("4. can hide Currency Field", () => {
-    const defaultValue = 1000;
-    // Add new custom field
-    addCustomField("Currency Input");
-    cy.testJsontext("defaultvalue", defaultValue);
-    hideAndVerifyProperties("customField1", defaultValue);
-    removeCustomField();
-  });
-
-  it("5. can hide Date Field", () => {
+  it("5. hides fields on first load", () => {
     cy.openPropertyPane("jsonformwidget");
-    cy.openFieldConfiguration("dob");
 
-    hideAndVerifyProperties("dob", "10/12/1992");
-  });
-
-  it("6. can hide Input Field", () => {
-    cy.openPropertyPane("jsonformwidget");
+    // hide education field
+    cy.openFieldConfiguration("education");
+    cy.togglebarDisable(".t--property-control-visible input");
+    cy.get(backBtn).click({ force: true }).wait(500);
+    // hide name field
     cy.openFieldConfiguration("name");
+    cy.togglebarDisable(".t--property-control-visible input");
 
-    hideAndVerifyProperties("name", "John");
-  });
+    // publish the app
+    cy.PublishtheApp();
+    cy.wait(4000);
 
-  it("7. can hide Multiselect Field", () => {
-    cy.openPropertyPane("jsonformwidget");
-    cy.openFieldConfiguration("hobbies");
+    // Check if name is hidden
+    cy.get(`${fieldPrefix}-name`).should("not.exist");
+    // Check if education is hidden
+    cy.get(`${fieldPrefix}-education`).should("not.exist");
 
-    hideAndVerifyProperties("hobbies", ["travelling", "swimming"]);
-  });
+    // check if name and education are not present in form data
+    cy.get(".t--widget-textwidget .bp3-ui-text").then(($el) => {
+      const formData = JSON.parse($el.text());
 
-  it("8. can hide Object Field", () => {
-    cy.openPropertyPane("jsonformwidget");
-    cy.openFieldConfiguration("address");
-
-    hideAndVerifyProperties("address", {
-      street: "Koramangala",
-      city: "Bangalore",
+      cy.wrap(formData.name).should("deep.equal", undefined);
+      cy.wrap(formData.education).should("deep.equal", undefined);
     });
   });
 });
