+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/RTE/RichTextEditor_spec.js
@@ -38,7 +38,6 @@ describe("RichTextEditor Widget Functionality", function () {
   });
 
   beforeEach(() => {
-    cy.wait(7000);
     cy.openPropertyPane("richtexteditorwidget");
   });
 
@@ -195,124 +194,6 @@ describe("RichTextEditor Widget Functionality", function () {
     cy.openPropertyPane("richtexteditorwidget");
     cy.testJsontext("defaultvalue", "c");
     cy.get(".t--widget-textwidget").should("contain", "false");
-  });
-
-  it("7. Check if the binding is getting removed from the text and the RTE widget", function () {
-    cy.openPropertyPane("textwidget");
-    cy.updateCodeInput(".t--property-control-text", `{{RichtextEditor.text}}`);
-    // Change defaultText of the RTE
-    cy.openPropertyPane("richtexteditorwidget");
-    cy.testJsontext("defaultvalue", "Test Content");
-
-    //Check if the text widget has the defaultText of RTE
-    cy.get(".t--widget-textwidget").should("contain", "Test Content");
-
-    //Clear the default text from RTE
-    cy.openPropertyPane("richtexteditorwidget");
-    cy.testJsontext("defaultvalue", "");
-
-    //Check if text widget and RTE widget does not have any text in it.
-    cy.get(".t--widget-richtexteditorwidget").should("contain", "");
-    cy.get(".t--widget-textwidget").should("contain", "");
-  });
-
-  it("8. Check if text does not re-appear when cut, inside the RTE widget", function () {
-    cy.window().then((win) => {
-      const tinyMceId = "rte-6h8j08u7ea";
-
-      const editor = win.tinymce.editors[tinyMceId];
-
-      //Set the content
-      editor.setContent("Test Content");
-
-      //Check the content:
-      expect(editor.getContent({ format: "text" })).to.be.equal("Test Content");
-
-      //Set the content
-      editor.setContent("");
-
-      //Check the content:
-      expect(editor.getContent({ format: "text" })).to.be.equal("");
-    });
-  });
-
-  it("9. Check if the cursor position is at the end for the RTE widget", function () {
-    const tinyMceId = "rte-6h8j08u7ea";
-    const testString = "Test Content";
-    const testStringLen = testString.length;
-
-    // Check if the cursor is at the end when input Type is HTML
-    setRTEContent(testString);
-    testCursorPoistion(testStringLen, tinyMceId);
-    setRTEContent("{selectAll}");
-    setRTEContent("{backspace}");
-
-    // Changing the input type to markdown and again testing the cursor position
-    cy.openPropertyPane("richtexteditorwidget");
-    cy.get("span:contains('Markdown')").eq(0).click({ force: true });
-    setRTEContent(testString);
-    testCursorPoistion(testStringLen, tinyMceId);
-  });
-
-  it("10. Check if different font size texts are supported inside the RTE widget", function () {
-    const tinyMceId = "rte-6h8j08u7ea";
-    const testString = "Test Content";
-
-    // Set the content inside RTE widget by typing
-    setRTEContent(`${testString} {enter} ${testString} 1`);
-
-    cy.get(".tox-tbtn--bespoke").click({ force: true });
-    cy.contains("Heading 1").click({ force: true });
-
-    cy.window().then((win) => {
-      const editor = win.tinymce.editors[tinyMceId];
-
-      // Get the current editor text
-      const getCurrentHtmlContent = editor.getContent();
-
-      // Check if the editor contains text of font sizes h1 and p;
-      expect(getCurrentHtmlContent).contains("<h1>");
-      expect(getCurrentHtmlContent).contains("<p>");
-    });
-  });
-
-  it("11. Check if button for Underline exists within the Toolbar of RTE widget", () => {
-    cy.get('[aria-label="Underline"]').should("exist");
-
-    //Check if button for Background Color is rendered only once within the Toolbar of RTE widget
-    cy.get('[aria-label="Background color"]').should("have.length", 1);
-
-    //Check if button for Text Color is rendered only once within the Toolbar of RTE widget
-    cy.get('[aria-label="Text color"]').should("have.length", 1);
-  });
-
-  it("12. Check if able to add an emoji through toolbar", () => {
-    cy.get('[aria-label="More..."]').click({ force: true });
-    cy.wait(500);
-    cy.get('[aria-label="Emoticons"]').click({ force: true });
-    cy.wait(500);
-    cy.get('[aria-label="grinning"]').click({ force: true });
-    const getEditorContent = (win) => {
-      const tinyMceId = "rte-6h8j08u7ea";
-      const editor = win.tinymce.editors[tinyMceId];
-      return editor.getContent();
-    };
-
-    //contains emoji
-    cy.window().then((win) => {
-      expect(getEditorContent(win)).contains("😀");
-    });
-
-    //trigger a backspace
-    cy.get(formWidgetsPage.richTextEditorWidget + " iframe").then(($iframe) => {
-      const $body = $iframe.contents().find("body");
-      cy.get($body).type("{backspace}", { force: true });
-    });
-
-    // after backspace the emoji should not be present
-    cy.window().then((win) => {
-      expect(getEditorContent(win)).not.contains("😀");
-    });
   });
 
   afterEach(() => {
