+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/Tab/Tab_spec.js
@@ -158,17 +158,11 @@ describe("Tab widget test", function () {
   });
   it("11. Validates Total Number Of Tabs Displayed In The Property Pane", () => {
     cy.get(Layoutpage.tabNumber).should("have.text", "2");
-  });
-  it("12. Validates Total Number Of Tabs Displayed In The Property Pane After Adding A Tab", () => {
+    //Validates Total Number Of Tabs Displayed In The Property Pane After Adding A Tab
     cy.get(Layoutpage.tabButton).last().click({ force: true });
     cy.get(Layoutpage.tabNumber).should("have.text", "3");
-  });
-  it("13. Validates Total Number Of Tabs Displayed In The Property Pane After Deleting A Tab", () => {
+    //Validates Total Number Of Tabs Displayed In The Property Pane After Deleting A Tab
     cy.get(Layoutpage.tabDelete).eq(1).click({ force: true });
     cy.get(Layoutpage.tabNumber).should("have.text", "2");
   });
-});
-
-afterEach(() => {
-  // put your clean up code if any
 });
