+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/TableV1/TableFilter2_Spec.ts
@@ -257,156 +257,4 @@ describe("Verify various Table_Filter combinations", function () {
     });
     table.RemoveFilterNVerify("2381224", true, false);
   });
-
-  it("9. Verify Full table data - download csv and download Excel", function () {
-    table.DownloadFromTable("Download as CSV");
-    //This plugin works only from cypress ^9.2
-    //cy.verifyDownload("Table1.csv")
-    table.ValidateDownloadNVerify("Table1.csv", "Michael Lawson");
-
-    table.DownloadFromTable("Download as Excel");
-    table.ValidateDownloadNVerify("Table1.xlsx", "Michael Lawson");
-  });
-
-  it("10. Verify Searched data - download csv and download Excel", function () {
-    table.SearchTable("7434532");
-    table.ReadTableRowColumnData(0, 3).then((afterSearch) => {
-      expect(afterSearch).to.eq("Byron Fields");
-    });
-
-    table.DownloadFromTable("Download as CSV");
-    //This plugin works only from cypress ^9.2
-    //cy.verifyDownload("Table1.csv")
-    table.ValidateDownloadNVerify("Table1.csv", "byron.fields@reqres.in");
-
-    table.DownloadFromTable("Download as Excel");
-    table.ValidateDownloadNVerify("Table1.xlsx", "Ryan Holmes");
-
-    table.RemoveSearchTextNVerify("2381224");
-
-    table.DownloadFromTable("Download as CSV");
-    table.ValidateDownloadNVerify("Table1.csv", "2736212");
-
-    table.DownloadFromTable("Download as Excel");
-    table.ValidateDownloadNVerify("Table1.xlsx", "Beef steak");
-  });
-
-  it("11. Verify Filtered data - download csv and download Excel", function () {
-    table.OpenNFilterTable("id", "starts with", "6");
-    table.ReadTableRowColumnData(0, 3).then(($cellData) => {
-      expect($cellData).to.eq("Tobias Funke");
-    });
-    table.CloseFilter();
-
-    table.DownloadFromTable("Download as CSV");
-    //This plugin works only from cypress ^9.2
-    //cy.verifyDownload("Table1.csv")
-    table.ValidateDownloadNVerify("Table1.csv", "Beef steak");
-
-    table.DownloadFromTable("Download as Excel");
-    table.ValidateDownloadNVerify("Table1.xlsx", "tobias.funke@reqres.in");
-
-    agHelper.GetNClick(table._filterBtn);
-    table.RemoveFilterNVerify("2381224", true, false);
-
-    table.DownloadFromTable("Download as CSV");
-    table.ValidateDownloadNVerify("Table1.csv", "Tuna Salad");
-
-    table.DownloadFromTable("Download as Excel");
-    table.ValidateDownloadNVerify("Table1.xlsx", "Avocado Panini");
-  });
-
-  it("12. Import TableFilter application & verify all filters for same FirstName (one word column) + Bug 13334", () => {
-    deployMode.NavigateBacktoEditor();
-    table.WaitUntilTableLoad();
-    homePage.NavigateToHome();
-    homePage.ImportApp("TableFilterImportApp.json");
-    homePage.AssertImportToast();
-    deployMode.DeployApp();
-    table.WaitUntilTableLoad();
-
-    //Contains
-    table.OpenNFilterTable("FirstName", "contains", "Della");
-    table.ReadTableRowColumnData(0, 3).then(($cellData) => {
-      expect($cellData).to.eq("Alvarado");
-    });
-
-    filterOnlyCondition("does not contain", "49");
-    filterOnlyCondition("starts with", "1");
-
-    //Ends with - Open Bug 13334
-    filterOnlyCondition("ends with", "1");
-
-    filterOnlyCondition("is exactly", "1");
-    filterOnlyCondition("empty", "0");
-    filterOnlyCondition("not empty", "50");
-    filterOnlyCondition("starts with", "3", "ge");
-    table.ReadTableRowColumnData(0, 3).then(($cellData) => {
-      expect($cellData).to.eq("Chandler");
-    });
-
-    table.OpenNFilterTable("FullName", "ends with", "ross", "OR", 1);
-    agHelper
-      .GetText(table._showPageItemsCount)
-      .then(($count) => expect($count).contain("4"));
-    table.CloseFilter();
-    agHelper
-      .GetText(table._filtersCount)
-      .then(($count) => expect($count).contain("2"));
-
-    table.OpenFilter();
-    table.RemoveFilterNVerify("1", true, false);
-  });
-
-  it("13. Verify all filters for same FullName (two word column) + Bug 13334", () => {
-    //Contains
-    table.OpenNFilterTable("FullName", "contains", "torres");
-    table.ReadTableRowColumnData(0, 2).then(($cellData) => {
-      expect($cellData).to.eq("Virgie");
-    });
-
-    filterOnlyCondition("does not contain", "49");
-    filterOnlyCondition("starts with", "0");
-    filterOnlyCondition("ends with", "1");
-    filterOnlyCondition("is exactly", "0");
-    filterOnlyCondition("empty", "0");
-    filterOnlyCondition("not empty", "50");
-    filterOnlyCondition("contains", "1", "wolf");
-    table.ReadTableRowColumnData(0, 2).then(($cellData) => {
-      expect($cellData).to.eq("Teresa");
-    });
-
-    table.OpenNFilterTable("FirstName", "starts with", "wa", "OR", 1);
-    agHelper.Sleep();
-    agHelper
-      .GetText(table._showPageItemsCount)
-      .then(($count) => expect($count).contain("3"));
-
-    table.OpenNFilterTable("LastName", "ends with", "son", "OR", 2);
-    agHelper
-      .GetText(table._showPageItemsCount)
-      .then(($count) => expect($count).contain("10"));
-    table.CloseFilter();
-    agHelper
-      .GetText(table._filtersCount)
-      .then(($count) => expect($count).contain("3"));
-
-    table.OpenFilter();
-    table.RemoveFilterNVerify("1", true, false);
-  });
-
-  function filterOnlyCondition(
-    condition: string,
-    expectedCount: string,
-    input: string | "" = "",
-  ) {
-    agHelper.GetNClick(table._filterConditionDropdown);
-    cy.get(table._dropdownText).contains(condition).click();
-    if (input)
-      agHelper.GetNClick(table._filterInputValue, 0).type(input).wait(500);
-    agHelper.ClickButton("APPLY");
-    agHelper
-      .GetText(table._showPageItemsCount)
-      .then(($count) => expect($count).contain(expectedCount));
-  }
 });
