+++ app/client/cypress/e2e/Regression/ClientSide/Widgets/TableV2/Inline_editing_spec.js
@@ -299,488 +299,4 @@ describe("Table widget inline editing functionality", () => {
     cy.get(".t--property-control-editable .ads-v2-switch").click();
     cy.get("[data-rbd-draggable-id='EditActions1']").should("not.exist");
   });
-
-  it("11. should check that save/discard column is added/removed when inline save option is changed", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("exist");
-    cy.get(".t--property-control-updatemode .t--property-control-label")
-      .last()
-      .click();
-    cy.get(".ads-v2-segmented-control-value-CUSTOM").click({ force: true });
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("not.exist");
-    cy.makeColumnEditable("task");
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("not.exist");
-    cy.get(".t--property-control-updatemode .t--property-control-label")
-      .last()
-      .click();
-    cy.get(".ads-v2-segmented-control-value-ROW_LEVEL").click({ force: true });
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("exist");
-    cy.get(".t--property-control-updatemode .t--property-control-label")
-      .last()
-      .click();
-    cy.get(".ads-v2-segmented-control-value-CUSTOM").click({ force: true });
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("not.exist");
-    cy.makeColumnEditable("step");
-    cy.makeColumnEditable("task");
-    cy.get(".t--property-control-updatemode .t--property-control-label")
-      .last()
-      .click();
-    cy.get(".ads-v2-segmented-control-value-ROW_LEVEL").click({ force: true });
-    cy.get("[data-rbd-draggable-id='EditActions1']").should("not.exist");
-  });
-
-  it("12. should check that cell of an editable column is editable", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    // click the edit icon
-    cy.editTableCell(0, 0);
-    cy.get(
-      "[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input",
-    ).should("not.be.disabled");
-
-    //double click the cell
-    cy.openPropertyPane("tablewidgetv2");
-    cy.get(
-      `[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input`,
-    ).should("not.exist");
-    cy.get(`[data-colindex=0][data-rowindex=0] .t--table-text-cell`).trigger(
-      "dblclick",
-    );
-    cy.get(
-      `[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input`,
-    ).should("exist");
-    cy.get(
-      "[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input",
-    ).should("not.be.disabled");
-  });
-
-  it("13. should check that changes can be discarded by clicking escape", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    let value;
-    cy.readTableV2data(0, 0).then((val) => {
-      value = val;
-    });
-    cy.makeColumnEditable("step");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "newValue");
-    cy.discardTableCellValue(0, 0);
-    cy.get(
-      `[data-colindex="0"][data-rowindex="0"] .t--inlined-cell-editor input.bp3-input`,
-    ).should("not.exist");
-    cy.readTableV2data(0, 0).then((val) => {
-      expect(val).to.equal(value);
-    });
-  });
-
-  it("14. should check that changes can be saved by pressing enter or clicking outside", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    let value;
-    cy.readTableV2data(0, 0).then((val) => {
-      value = val;
-    });
-    cy.makeColumnEditable("step");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "newValue");
-    cy.saveTableCellValue(0, 0);
-    cy.get(
-      `[data-colindex="0"][data-rowindex="0"] .t--inlined-cell-editor input.bp3-input`,
-    ).should("not.exist");
-    cy.wait(1000);
-    cy.readTableV2data(0, 0).then((val) => {
-      expect(val).to.not.equal(value);
-      value = val;
-    });
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "someOtherNewValue");
-    cy.openPropertyPane("tablewidgetv2");
-    cy.get(
-      `[data-colindex="0"][data-rowindex="0"] .t--inlined-cell-editor input.bp3-input`,
-    ).should("not.exist");
-    cy.wait(1000);
-    cy.readTableV2data(0, 0).then((val) => {
-      expect(val).to.not.equal(value);
-    });
-  });
-
-  it("15. should check that updatedRows and updatedRowIndices have correct values", () => {
-    cy.dragAndDropToCanvas("textwidget", { x: 300, y: 500 });
-    cy.openPropertyPane("textwidget");
-    cy.updateCodeInput(".t--property-control-text", `{{Table1.updatedRows}}`);
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "newValue");
-    cy.saveTableCellValue(0, 0);
-    cy.get(".t--widget-textwidget .bp3-ui-text").should(
-      "contain",
-      `[  {    "index": 0,    "updatedFields": {      "step": "newValue"    },    "allFields": {      "step": "newValue",      "task": "Drop a table",      "status": "✅"    }  }]`,
-    );
-    cy.openPropertyPane("textwidget");
-    cy.updateCodeInput(
-      ".t--property-control-text",
-      `{{Table1.updatedRowIndices}}`,
-    );
-    cy.get(".t--widget-textwidget .bp3-ui-text").should("contain", "[  0]");
-  });
-
-  it("16. should check that onsubmit event is available for the columns that are editable", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.editColumn("step");
-    [
-      {
-        columnType: "URL",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Number",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Date",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Image",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Video",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Menu button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Icon button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Plain text",
-        expected: "not.exist",
-      },
-    ].forEach((data) => {
-      cy.get(commonlocators.changeColType).last().click();
-      cy.get(".t--dropdown-option")
-        .children()
-        .contains(data.columnType)
-        .click();
-      cy.wait("@updateLayout");
-      cy.wait(500);
-      cy.get(".t--property-control-onsubmit").should(data.expected);
-    });
-
-    cy.get(propPaneBack).click();
-    cy.makeColumnEditable("step");
-    cy.editColumn("step");
-
-    [
-      {
-        columnType: "URL",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Number",
-        expected: "exist",
-      },
-      {
-        columnType: "Date",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Image",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Video",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Menu button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Icon button",
-        expected: "not.exist",
-      },
-      {
-        columnType: "Plain text",
-        expected: "exist",
-      },
-    ].forEach((data) => {
-      cy.get(commonlocators.changeColType).last().click();
-      cy.get(".t--dropdown-option")
-        .children()
-        .contains(data.columnType)
-        .click();
-      cy.wait("@updateLayout");
-      cy.wait(500);
-      cy.get(".t--property-control-onsubmit").should(data.expected);
-    });
-  });
-
-  it("17. should check that onsubmit event is triggered when changes are saved", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("step");
-    cy.getAlert("onSubmit", "Submitted!!");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "NewValue");
-    cy.saveTableCellValue(0, 0);
-
-    cy.get(widgetsPage.toastAction).should("be.visible");
-    cy.get(widgetsPage.toastActionText)
-      .last()
-      .invoke("text")
-      .then((text) => {
-        expect(text).to.equal("Submitted!!");
-      });
-  });
-
-  it("18. should check that onSubmit events has access to edit values through triggeredRow", () => {
-    const value = "newCellValue";
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("step");
-    cy.getAlert("onSubmit", "{{Table1.triggeredRow.step}}");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, value);
-    cy.saveTableCellValue(0, 0);
-
-    cy.get(widgetsPage.toastAction).should("be.visible");
-    cy.get(widgetsPage.toastActionText)
-      .last()
-      .invoke("text")
-      .then((text) => {
-        expect(text).to.equal(value);
-      });
-  });
-
-  it("19. should check that onSave is working", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("EditActions1");
-    //cy.get(".t--property-pane-section-collapse-savebutton").click({force:true});
-    cy.get(".t--property-pane-section-collapse-discardbutton").click({
-      force: true,
-    });
-    cy.getAlert("onSave", "Saved!!");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "NewValue");
-    cy.openPropertyPane("tablewidgetv2");
-    cy.saveTableRow(4, 0);
-    cy.get(widgetsPage.toastAction).should("be.visible");
-    cy.get(widgetsPage.toastActionText)
-      .last()
-      .invoke("text")
-      .then((text) => {
-        expect(text).to.equal("Saved!!");
-      });
-  });
-
-  it("20. should check that onSave events has access to edit values through triggeredRow", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("EditActions1");
-    //cy.get(".t--property-pane-section-collapse-savebutton").click({force:true});
-    cy.get(".t--property-pane-section-collapse-discardbutton").click({
-      force: true,
-    });
-    cy.getAlert("onSave", "{{Table1.triggeredRow.step}}");
-    /*
-    cy.addSuccessMessage(
-      "{{Table1.triggeredRow.step}}",
-      ".t--property-control-onsave",
-    );
-    */
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "NewValue");
-    cy.openPropertyPane("tablewidgetv2");
-    cy.saveTableRow(4, 0);
-    cy.get(widgetsPage.toastAction).should("be.visible");
-    cy.get(widgetsPage.toastActionText)
-      .last()
-      .invoke("text")
-      .then((text) => {
-        expect(text).to.equal("NewValue");
-      });
-  });
-
-  it("21. should check that onDiscard event is working", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("EditActions1");
-    cy.get(".t--property-pane-section-collapse-savebutton").click();
-    //cy.get(".t--property-pane-section-collapse-discardbutton").click();
-    cy.getAlert("onDiscard", "discarded!!");
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "NewValue");
-    cy.openPropertyPane("tablewidgetv2");
-    cy.discardTableRow(4, 0);
-    cy.get(widgetsPage.toastAction).should("be.visible");
-    cy.get(widgetsPage.toastActionText)
-      .last()
-      .invoke("text")
-      .then((text) => {
-        expect(text).to.equal("discarded!!");
-      });
-  });
-
-  it("22. should check that inline editing works with text wrapping disabled", () => {
-    cy.addDsl(dsl);
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editTableCell(0, 0);
-    cy.get(
-      "[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input",
-    ).should("not.be.disabled");
-  });
-
-  it("23. should check that inline editing works with text wrapping enabled", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("step");
-    cy.get(".t--property-control-cellwrapping .ads-v2-switch")
-      .first()
-      .click({ force: true });
-    cy.editTableCell(0, 0);
-    cy.get(
-      "[data-colindex=0][data-rowindex=0] .t--inlined-cell-editor input.bp3-input",
-    ).should("not.be.disabled");
-  });
-
-  it("24. should check that doesn't grow taller when text wrapping is disabled", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editTableCell(0, 0);
-    cy.get(
-      "[data-colindex='0'][data-rowindex='0'] .t--inlined-cell-editor",
-    ).should("have.css", "height", "32px");
-    cy.enterTableCellValue(0, 0, "this is a very long cell value");
-    cy.get(
-      "[data-colindex='0'][data-rowindex='0'] .t--inlined-cell-editor",
-    ).should("have.css", "height", "32px");
-  });
-
-  it("25. should check that grows taller when text wrapping is enabled", () => {
-    cy.openPropertyPane("tablewidgetv2");
-    cy.makeColumnEditable("step");
-    cy.editColumn("step");
-    cy.get(".t--property-control-cellwrapping .ads-v2-switch").click({
-      force: true,
-    });
-    cy.editTableCell(0, 0);
-    cy.get(
-      "[data-colindex='0'][data-rowindex='0'] .t--inlined-cell-editor",
-    ).should("have.css", "height", "34px");
-    cy.enterTableCellValue(0, 0, "this is a very long cell value");
-    cy.get(
-      "[data-colindex='0'][data-rowindex='0'] .t--inlined-cell-editor",
-    ).should("not.have.css", "height", "34px");
-  });
-
-  it("26. should check if updatedRowIndex is getting updated for single row update mode", () => {
-    cy.dragAndDropToCanvas("textwidget", { x: 400, y: 400 });
-    cy.get(".t--widget-textwidget").should("exist");
-    cy.updateCodeInput(
-      ".t--property-control-text",
-      `{{Table1.updatedRowIndex}}`,
-    );
-
-    cy.dragAndDropToCanvas("buttonwidget", { x: 300, y: 300 });
-    cy.get(".t--widget-buttonwidget").should("exist");
-    cy.get(PROPERTY_SELECTOR.onClick).find(".t--js-toggle").click();
-    cy.updateCodeInput(".t--property-control-label", "Reset");
-    cy.updateCodeInput(
-      PROPERTY_SELECTOR.onClick,
-      `{{resetWidget("Table1",true)}}`,
-    );
-
-    // case 1: check if updatedRowIndex has -1 as the default value:
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-
-    cy.openPropertyPane("tablewidgetv2");
-
-    cy.makeColumnEditable("step");
-    cy.wait(1000);
-
-    // case 2: check if updatedRowIndex is 0, when cell at row 0 is updated.
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "#12").type("{enter}");
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", 0);
-
-    // case 3: check if updatedRowIndex is -1 when changes are discarded.
-    cy.discardTableRow(4, 0);
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-
-    // case 4: check if the updateRowIndex is -1 when widget is reset
-    cy.editTableCell(0, 1);
-    cy.enterTableCellValue(0, 1, "#13").type("{enter}");
-    cy.contains("Reset").click({ force: true });
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-
-    // case 5: check if the updatedRowIndex changes to -1 when the table data changes.
-    cy.wait(1000);
-    cy.editTableCell(0, 2);
-    cy.enterTableCellValue(0, 2, "#14").type("{enter}");
-    cy.openPropertyPane("tablewidgetv2");
-    cy.get(widgetsPage.tabedataField).type("{backspace}");
-    cy.wait(300);
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-  });
-
-  it("27. should check if updatedRowIndex is getting updated for multi row update mode", () => {
-    cy.dragAndDropToCanvas("textwidget", { x: 400, y: 400 });
-    cy.get(".t--widget-textwidget").should("exist");
-    cy.updateCodeInput(
-      ".t--property-control-text",
-      `{{Table1.updatedRowIndex}}`,
-    );
-
-    cy.dragAndDropToCanvas("buttonwidget", { x: 300, y: 300 });
-    cy.get(".t--widget-buttonwidget").should("exist");
-    cy.get(PROPERTY_SELECTOR.onClick).find(".t--js-toggle").click();
-    cy.updateCodeInput(".t--property-control-label", "Reset");
-    cy.updateCodeInput(
-      PROPERTY_SELECTOR.onClick,
-      `{{resetWidget("Table1",true)}}`,
-    );
-
-    cy.openPropertyPane("tablewidgetv2");
-
-    cy.makeColumnEditable("step");
-    cy.get(".ads-v2-segmented-control-value-CUSTOM").click({ force: true });
-    cy.wait(1000);
-
-    // case 1: check if updatedRowIndex is 0, when cell at row 0 is updated.
-    cy.editTableCell(0, 0);
-    cy.enterTableCellValue(0, 0, "#12").type("{enter}");
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", 0);
-
-    // case 2: check if the updateRowIndex is -1 when widget is reset
-    cy.editTableCell(0, 1);
-    cy.enterTableCellValue(0, 1, "#13").type("{enter}");
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", 1);
-    cy.contains("Reset").click({ force: true });
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-
-    // case 3: check if the updatedRowIndex changes to -1 when the table data changes.
-    cy.wait(1000);
-    cy.editTableCell(0, 2);
-    cy.enterTableCellValue(0, 2, "#14").type("{enter}");
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", 2);
-    cy.openPropertyPane("tablewidgetv2");
-    cy.get(widgetsPage.tabedataField).type("{backspace}");
-    cy.wait(300);
-    cy.get(commonlocators.textWidgetContainer).should("contain.text", -1);
-  });
 });
